iter12: Изменена структура пакета middleware
