--- internal/gzip/reader.go
@@ -1,34 +1,0 @@
-package gzip
-
-import (
-	"compress/gzip"
-	"io"
-)
-
-type CompressReader struct {
-	r  io.ReadCloser
-	zr *gzip.Reader
-}
-
-func NewCompressReader(r io.ReadCloser) (*CompressReader, error) {
-	zr, err := gzip.NewReader(r)
-	if err != nil {
-		return nil, err
-	}
-
-	return &CompressReader{
-		r:  r,
-		zr: zr,
-	}, nil
-}
-
-func (c CompressReader) Read(p []byte) (n int, err error) {
-	return c.zr.Read(p)
-}
-
-func (c *CompressReader) Close() error {
-	if err := c.r.Close(); err != nil {
-		return err
-	}
-	return c.zr.Close()
-}
